backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
pydantic
python-dotenv
apscheduler
